requirements.txt
asttokens==2.2.1
backcall==0.2.0
Bottleneck==1.3.7
certifi==2023.7.22
charset-normalizer==3.2.0
clickhouse-driver==0.2.6
comm==0.1.3
# data-wrapper==2.2.2
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
et-xmlfile==1.1.0
executing==1.2.0
idna==3.4
ipykernel==6.25.0
ipython==8.14.0
jedi==0.19.0
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numexpr==2.8.4
numpy==1.25.1
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prettytable==3.8.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
scikit-learn==1.1.3
scipy==1.11.1
six==1.16.0
stack-data==0.6.2
threadpoolctl==3.2.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
wcwidth==0.2.6
wrapt==1.15.0
